requirements.txt
flask
pandas
google-genai
python-dotenv
psycopg2-binary
